Backend/localization_service/requirements.txt
# Localization Service Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
scipy==1.12.0
scikit-learn==1.3.2
pydantic==2.5.0
aiofiles==23.2.1
prometheus_client==0.19.0

# SLAM Integration
pybind11==2.11.1
pyyaml==6.0.1
matplotlib==3.8.3
pillow==10.2.0

# Production SLAM Systems (install via Docker)
# stella-vslam  # Installed in Docker container
# pycolmap==0.4.0  # For COLMAP integration (installed in Docker)

# For production SLAM support
requests==2.31.0
toml==0.10.2
imageio==2.31.5
open3d==0.18.0  # For 3D visualization and processing

# Real-time processing
websockets==12.0
asyncio-mqtt==0.16.1
aiohttp==3.9.1

# VIO Integration
# scipy==1.11.4  # Already included above
collections-extended==2.0.2  # For advanced data structures
filterpy==1.4.5  # For Kalman filtering
numba==0.59.0  # For performance optimization

# Development and debugging
python-multipart==0.0.6

# Enterprise OpenTelemetry Dependencies - Critical 60fps Performance Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-aiohttp-client==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-propagator-b3==1.21.0
opentelemetry-propagator-jaeger==1.21.0